requirement.txt
graia-application-mirai
graia-template
loguru
json5
pillow
baidu-aip
numpy
pyyaml
tensorflow
scikit-image